+++ services/web/server/tests/unit/with_dbs/02/conftest.py
@@ -10,7 +10,6 @@ from copy import deepcopy
 from pathlib import Path
 from typing import Any, AsyncIterator, Awaitable, Callable
 from unittest import mock
-from unittest.mock import Mock
 
 import pytest
 from aiohttp import web
@@ -349,36 +348,3 @@ def mock_catalog_service_api_responses(client, aioresponses_mocker):
         url_pattern,
         repeat=True,
     )
-
-
-@pytest.fixture
-def mock_rabbitmq(mocker: MockerFixture) -> None:
-    mocker.patch(
-        "simcore_service_webserver.director_v2_core_dynamic_services.get_rabbitmq_client",
-        autospec=True,
-    )
-
-
-@pytest.fixture
-def mock_progress_bar(mocker: MockerFixture) -> None:
-
-    sub_progress = Mock()
-
-    class MockedProgress:
-        async def __aenter__(self):
-            return self
-
-        async def __aexit__(self, *args):
-            pass
-
-        def sub_progress(self, *kwargs):  # pylint:disable=no-self-use
-            return sub_progress
-
-    mock_bar = MockedProgress()
-
-    mocker.patch(
-        "simcore_service_webserver.director_v2_core_dynamic_services.ProgressBarData",
-        autospec=True,
-        return_value=mock_bar,
-    )
-    return mock_bar
